% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqvis.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean NOAA Earthquake Data}
\usage{
eq_clean_data(data)
}
\arguments{
\item{data}{\code{data.frame} of NOAA significant earthquake dataset}
}
\value{
Returns a cleaned \code{data.frame}. The country and colon are removed from the LOCATION_NAME 
and it is converted to title case. A new \code{DATE} column is derived from the date parts (\code{YEAR, MONTH, DAY}). \code{LATITUDE} 
and \code{LONGITUDE} are converted to numeric.
}
\description{
This function cleans an NOAA earthquake significant earthquake \code{data.frame}.
}
\details{
If \code{MONTH} or \code{DAY} are missing, 1 is used (January or 1st of month, respectively).
}
\note{
The NOAA earthquake dataset can be accessed at: 
\url{https://www.ngdc.noaa.gov/nndc/struts/results?type_0=Exact&query_0=$ID&t=101650&s=13&d=189&dfn=signif.txt}.
This dataset contains information about 5,933 earthquakes around the world over an approximately 4,000 year time span. 
This package was developed to work with years AC only.
}
\examples{
\dontrun{eq_clean_data(sample_eq)}

}
