% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqvis.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create NOAA Earthquake Label}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{\code{data.frame} of cleaned (see \code{eq_clean_data}) NOAA significant earthquake dataset}
}
\value{
Returns a label vector of popup text including location, magnitude, and deaths.
}
\description{
This function creates a label vector of location, magnitude and deaths for display on a map circle marker popup.
}
\details{
If \code{LOCATION_NAME}, \code{EQ_PRIMARY} or \code{DEATHS} are \code{NA}, they are not included in the label text.
}
\examples{
\dontrun{eq_create_label(sample_eq)}

}
