% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean_data.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Cleaning location column}
\usage{
eq_location_clean(location_name)
}
\arguments{
\item{location_name}{Column containing location data.}
}
\value{
Returns with a cleaned up \code{LOCATION_NAME} column with no country and colon present.
}
\description{
Removes the country and colon from all the entries in \code{LOCATION_NAME}
}
