% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqvis.R
\name{eq_map}
\alias{eq_map}
\title{Map NOAA Earthquake Data}
\usage{
eq_map(data, annot_col = "DATE")
}
\arguments{
\item{data}{\code{data.frame} of cleaned (see \code{eq_clean_data}) NOAA significant earthquake dataset}

\item{annot_col}{default 'DATE'; character vector to use for circle marker popup; See \code{eq_create_label} to create 
text label including location, magnitude, and deaths}
}
\value{
Returns a Leaflet map widget of the earthquake data.
}
\description{
This function creates a leaflet map of the significant earthquake data provided and includes a popup showing 
the value of the character vector provided. The size of the circle marker is based on the magnitude of the earthquake.
}
\examples{
\dontrun{sample_eq \%>\% 
eq_clean_data() \%>\% 
dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\% 
dplyr::mutate(popup_text = eq_create_label(.)) \%>\% 
eq_map(annot_col = "popup_text")}

}

