% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_read_raw.R
\docType{data}
\name{file_url}
\alias{file_url}
\title{Reading Raw NOAA Data}
\format{An object of class \code{character} of length 1.}
\usage{
file_url
}
\description{
Takes the raw data from the url and converts it into a tab delim data frame
}
\keyword{datasets}
