% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{Building a Timeline geom}
\format{An object of class \code{GeoTimeline} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomTimeline
}
\value{
Returns a timeline plot in date order of each earthquake that occur within certain countries and hows the
magnitude of each earthquake by its Richter Scale value as well as the number of deaths associated with each earthquake.
}
\description{
When running \code{geom_timeline} function, it produces a timeline of the
earthquakes to visualize them by date.
}
\examples{
\dontrun{
ggplot(data=subset(cleaned_data, !is.na(EQ_PRIMARY) & COUNTRY \%in\% c("RUSSIA")),
aes(x = DATE, y = COUNTRY, size = EQ_PRIMARY, xmin = as.Date('2000-01-01',"\%Y-\%m-\%d"),
   xmax = as.Date('2016-12-31',"\%Y-\%m-\%d"), color = DEATHS, fill = DEATHS, label = LOCATION_NAME)) +
 geom_timeline(alpha = .5) +
 scale_size_continuous(name = "Richter Scale Value") +
 scale_color_continuous(name = "# Deaths") +
 scale_fill_continuous(guide=FALSE) +
 labs(y = "") +
 theme(legend.position = "bottom",
       legend.key = element_blank(), legend.box = "horizontal", panel.background = element_blank(),
       axis.line.x = element_line(size = .5, color = "black"), axis.ticks.x = element_line(size = .5))
}
}
\keyword{datasets}
