% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{Adds labels to the top earthquakes based on magnitude and user input}
\format{An object of class \code{GeoTimelineLabel} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomTimelineLabel
}
\description{
Adds labels to the top earthquakes based on magnitude and user input
}
\examples{
\dontrun{
ggplot(data=subset(cleaned_data, !is.na(EQ_PRIMARY) & COUNTRY \%in\% c("USA", "RUSSIA")),
aes(x = DATE, y = COUNTRY, size = EQ_PRIMARY, xmin = as.Date('2000-01-01',"\%Y-\%m-\%d"),
   xmax = as.Date('2016-12-31',"\%Y-\%m-\%d"), color = DEATHS, fill = DEATHS, label = LOCATION_NAME)) +
 geom_timeline(alpha = .5) +
 scale_size_continuous(name = "Richter Scale Value") +
 scale_color_continuous(name = "# Deaths") +
 scale_fill_continuous(guide=FALSE) +
 labs(y = "") +
 geom_timeline_label(aes(n_max = 5)) +
 theme(legend.position = "bottom",
       legend.key = element_blank(), legend.box = "horizontal", panel.background = element_blank(),
       axis.line.x = element_line(size = 1, color = "black"), axis.ticks.x = element_line(size = 1))
}

}
\keyword{datasets}
