% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\name{stat_timeline}
\alias{stat_timeline}
\title{Timeline Statistic, for NOAA Earthquake Dataset}
\usage{
stat_timeline(mapping = NULL, data = NULL, geom = "timeline",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link{aes}} or
\code{\link{aes_}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
   options:

   If \code{NULL}, the default, the data is inherited from the plot
   data as specified in the call to \code{\link{ggplot}}.

   A \code{data.frame}, or other object, will override the plot
   data. All objects will be fortified to produce a data frame. See
   \code{\link{fortify}} for which variables will be created.

   A \code{function} will be called with a single argument,
   the plot data. The return value must be a \code{data.frame.}, and
   will be used as the layer data.}

\item{geom}{Use to override the default connection between
`geom_timeline` and `stat_timeline`.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}

\item{...}{other arguments passed on to \code{\link{layer}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\description{
The timeline stat is used to subset the earthquake data for the specified data range 
(aesthetics \code{xmin} and \code{xmax}) and max n points (aesthetic \code{n_max}) based on
the size (magnitude) aesthetic.
}
\section{Aesthetics}{

\Sexpr[results=rd,stage=build]{eqvis:::rd_aesthetics("StatTimeline","stat_timeline")}
}

\examples{
\dontrun{
ggplot(data=subset(eq_clean_data(sample_eq), !is.na(EQ_PRIMARY) & COUNTRY \%in\% c("USA","MEXICO")), 
   aes(x = DATE, y = COUNTRY, size = EQ_PRIMARY, xmin = as.Date('2000-01-01',"\%Y-\%m-\%d"), 
   xmax = as.Date('2016-12-31',"\%Y-\%m-\%d"), color = DEATHS, fill = DEATHS, label = LOCATION_NAME)) +
stat_timeline(alpha = .5) +
scale_size_continuous(name = "Richter Scale Value") +
scale_color_continuous(name = "# Deaths") +
scale_fill_continuous(guide=FALSE) +
labs(y = "") +
theme(legend.position = "bottom",
   legend.key = element_blank(), legend.box = "horizontal", panel.background = element_blank(), 
   axis.line.x = element_line(size = 1, color = "black"), axis.ticks.x = element_line(size = 1))
}

}
