% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\docType{data}
\name{StatTimeline}
\alias{StatTimeline}
\title{Stat function to support GeomTimeline and GeomTimelineLabel}
\format{An object of class \code{StatTimeline} (inherits from \code{Stat}, \code{ggproto}) of length 4.}
\usage{
StatTimeline
}
\description{
Stat function to support GeomTimeline and GeomTimelineLabel
}
\keyword{datasets}
