% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Cleaning up the Raw data}
\usage{
eq_clean_data(x)
}
\arguments{
\item{x}{data from the earthquake dataset}
}
\description{
The \code{eq_clean_data} function takes the raw data frame data and cleans up the data.
Takes the (\code{YEAR, MONTH, and DAY}) column and converts them into a date format.
It also sends the \code{LOCATION_NAME} to be cleaned up.
}
\note{
Added 1's for dates with missing values.
}
\examples{
\dontrun{
cleaned_data <- eq_clean_data(raw_data)
}
}
