% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Added Info for Popup Text}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{cleaned filtered data}
}
\description{
This function adds on to the \code{eq_map} function to include more
information to the pop up, adding \code{LOCATION NAME}, \code{EQ_PRIMARY},
and \code{TOTAL_DEATHS}.
}
\examples{
\dontrun{
readr::read_delim("earthquakes.tsv.gz", delim = "\\t") \%>\%
eq_clean_data() \%>\%
 dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
 dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
 eq_map(annot_col = "popup_text")
}

}
