% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_map}
\alias{eq_map}
\title{Earthquake Map}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{the filtered cleaned data}

\item{annot_col}{user defined data, for example picking \code{DATE}}
}
\description{
Function takes the filtered data frame and produces an interactive map that shows each
quakes epicenter based on there \code{LATITUDE} and \code{LONGITUDE}. The size of the radius of each circle is
based on the earthquakes magnitude \code{EQ_PRIMARY}.
}
\examples{
\dontrun{
readr::read_delim("earthquakes.tsv.gz", delim = "\\t") \%>\%
eq_clean_data() \%>\%
 dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
 eq_map(annot_col = "popup_text")
}

}
