% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqvis.R
\docType{data}
\name{sample_eq}
\alias{sample_eq}
\title{Sample NOAA Significant Earthquake Data}
\format{An object of class \code{data.frame} with 63 rows and 9 columns.}
\source{
\url{https://www.ngdc.noaa.gov/nndc/struts/results?type_0=Exact&query_0=$ID&t=101650&s=13&d=189&dfn=signif.txt}
}
\usage{
sample_eq
}
\description{
A dataset containing a subset of the NOAA significant earthquakes.
}
\details{
\itemize{
  \item YEAR. 
  \item MONTH.
  \item DAY.
  \item LOCATION_NAME.
  \item COUNTRY. 
  \item LATITUDE.
  \item LONGITUDE.
  \item DEATHS.
  \item EQ_PRIMARY. Magnitude of earthquake. 
}
}
\keyword{datasets}
